requirements.txt
einops
flax
absl-py

# For T5 tokenizer
transformers
sentencepiece

# Just used for image resizing
torch
torchvision
PySimpleGUI==4.60.4
Pillow==9.5.0
numpy==1.23.1
